Rename the timestamp conversion module to `_datetime`
Every other conversion module is named after the wrapper type it
produces, not after the protobuf message it consumes: `_delivery_area`,
`_location`, `_bounds`. This one converts to and from `datetime`, so
`_timestamp` reads like the odd one out, and the tests already agree --
they have always lived in `tests/proto/test_datetime.py`.

Rename it and leave the contents alone. Nothing public moves; the
symbols are re-exported from `frequenz.client.common.proto` either way.
