+++ src/frequenz/client/common/proto/__init__.py
@@ -4,7 +4,7 @@
 """General utilities for converting common types to/from protobuf types."""
 
 from ._enum import enum_from_proto
-from ._timestamp import datetime_from_proto, datetime_to_proto
+from ._datetime import datetime_from_proto, datetime_to_proto
 
 __all__ = [
     "datetime_from_proto",
